requirements.txt
# RSP: Reasoning-Style Poisoning
# Requirements for experiments

# Core
openai>=1.0.0
python-dotenv>=1.0.0

# Retrieval
rank-bm25>=0.2.2

# Evaluation & Analysis
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Defense (PPL calculation & PIGuard)
transformers>=4.30.0
torch>=2.0.0

# Utilities
tqdm>=4.65.0
jsonlines>=3.1.0
pyyaml>=6.0
